Add initial implementation of the CLI application and configuration files
* **Add CLI application and configuration files**
  - Add `src/app.py` to implement the Flask application with routes for configuring backup, viewing logs, and running recovery.
  - Add `src/__init__.py` to initialize the `src` directory as a Python package.
  - Add `src/config/__init__.py` to initialize the `config` directory as a Python package.
  - Add `src/config/dicom.toml` to provide detailed configurations for each source and backup system.
  - Add `src/config/logging.toml` to provide logging configurations.
  - Add `src/config/environments.toml` to provide environment-specific configurations.

* **Add data source classes**
  - Add `src/data_sources/__init__.py` to define the base class `DataSource`.
  - Add `src/data_sources/aria.py` to implement the `ARIA` class based on the `DataSource` base class.
  - Add `src/data_sources/mim.py` to implement the `MIM` class based on the `DataSource` base class.
  - Add `src/data_sources/mosaiq.py` to implement the `Mosaiq` class based on the `DataSource` base class.

* **Add backup system classes**
  - Add `src/backup_systems/__init__.py` to define the base class `BackupSystem`.
  - Add `src/backup_systems/orthanc.py` to implement the `Orthanc` class based on the `BackupSystem` base class.

* **Add unit tests**
  - Add `tests/__init__.py` to initialize the `tests` directory as a Python package.
  - Add `tests/test_aria.py` to write unit tests for the `ARIA` class.
  - Add `tests/test_mim.py` to write unit tests for the `MIM` class.
  - Add `tests/test_mosaiq.py` to write unit tests for the `Mosaiq` class.
  - Add `tests/test_orthanc.py` to write unit tests for the `Orthanc` class.

* **Update README.md**
  - Add instructions for running the CLI application.

* **Move old files**
  - Rename `scu_find_git_v1.py` to `old/scu_find_git_v1.py`.
  - Rename `scu_move_git_v1.py` to `old/scu_move_git_v1.py`.
