+++ old/get_treatment_report_git_v1.py
