+++ src/__init__.py
@@ -1,0 +1,1 @@
+# Initialize the src directory as a Python package
