+++ src/backup_systems/__init__.py
@@ -1,0 +1,6 @@
+class BackupSystem:
+    def store(self):
+        raise NotImplementedError
+
+    def verify(self):
+        raise NotImplementedError
