+++ src/config/__init__.py
@@ -1,0 +1,1 @@
+# Initialize the config directory as a Python package
