+++ src/data_sources/__init__.py
@@ -1,0 +1,6 @@
+class DataSource:
+    def query(self):
+        raise NotImplementedError
+
+    def transfer(self):
+        raise NotImplementedError
