+++ tests/__init__.py
@@ -1,0 +1,1 @@
+# Initialize the tests directory as a Python package
